# api/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "api.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/demo.service.v1.Demo/MediaIDGet": {
            "get": {
                "summary": "/demo.service.v1.Demo/MediaIDGet",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.MediaIDResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "fakeID",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "timestamp",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/demo.service.v1.Demo/Ping": {
            "get": {
                "summary": "/demo.service.v1.Demo/Ping",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/demo.service.v1.Demo/SayHello": {
            "get": {
                "summary": "/demo.service.v1.Demo/SayHello",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        },
        "/kratos-demo/say_hello": {
            "get": {
                "summary": "/kratos-demo/say_hello",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.demo.service.v1.HelloResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "demo.service.v1.Demo"
                ]
            }
        }
    },
    "definitions": {
        ".demo.service.v1.HelloReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            },
            "required": [
                "name"
            ]
        },
        ".demo.service.v1.HelloResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        ".demo.service.v1.MediaIDReq": {
            "type": "object",
            "properties": {
                "fakeID": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        ".demo.service.v1.MediaIDResp": {
            "type": "object",
            "properties": {
                "MediaID": {
                    "type": "string"
                }
            }
        },
        ".google.protobuf.Empty": {
            "type": "object",
            "properties": {},
            "description": "A generic empty message that you can re-use to avoid defining duplicated\n empty messages in your APIs. A typical example is to use it as the request\n or the response type of an API method. For instance:\n\n     service Foo {\n       rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n     }\n\n The JSON representation for `Empty` is empty JSON object `{}`."
        }
    }
}